type: feat
body:
1. 가입승인시 완료된 미팅목록으로 이동하도록 수정
2. 완료된 목록과 진행중인 미팅목록으로 폼 분할
3. 미팅 상세 페이지 들어갈 시 로그인중이 아니거나 미팅장이 아닌경우 접근방지 페이지로 이동하도록 수정
4. 완료된 미팅을 확인하기 위한 status를 DTO 엔티티에 추가
5.완료상태로 표기하기 위한 서비스 로직 추가
6. meeting/ 링크를 meeting/list로 수정

footer: 가입승인링크가 여러개 나오는 문제 발생 추후 수정

+++ src/main/java/sm/school/Controller/MeetingController.java
@@ -27,7 +27,7 @@ public class MeetingController {
     private final MeetingProposerService meetingProposerService;
 
 
-    @GetMapping("/")
+    @GetMapping("/list")
     public String MeetingList(Model model) {
 
         List<Meeting> meeting = meetingService.findMeeting();
@@ -35,6 +35,16 @@ public class MeetingController {
         model.addAttribute("meetings", meeting);
 
         return "meeting/list";
+    }
+
+    @GetMapping("/completeList")
+    public String MeetingCompleteList(Model model) {
+
+        List<Meeting> meeting = meetingService.findMeeting();
+
+        model.addAttribute("meetings", meeting);
+
+        return "meeting/completeList";
     }
 
     @GetMapping("/create")
@@ -59,7 +69,7 @@ public class MeetingController {
 
         meetingService.createMeeting(meetingDTO);
 
-        return "redirect:/";
+        return "redirect:/meeting/list";
     }
 
     @GetMapping("/detail")
@@ -72,7 +82,11 @@ public class MeetingController {
     }
 
     @GetMapping("/update")
-    public String updateMeetingFrom(@RequestParam("id") Long id, Model model) {
+    public String updateMeetingFrom(@RequestParam("id") Long id, Model model, Authentication authentication) {
+
+        if (authentication == null) {
+            return "redirect:/member/login";
+        }
 
         MeetingDTO meetingDTO = meetingService.selectMeeting(id);
         model.addAttribute("meetingDTO", meetingDTO);
@@ -81,19 +95,26 @@ public class MeetingController {
     }
 
     @PostMapping("/update")
-    public String updateMeeting(@ModelAttribute("meetingDTO") MeetingDTO meetingDTO) {
+    public String updateMeeting(@ModelAttribute("meetingDTO") MeetingDTO meetingDTO, Authentication authentication) {
+        if (authentication == null) {
+            return "redirect:/member/login";
+        }
 
         meetingService.updateMeeting(meetingDTO);
 
-        return "redirect:/meeting/";
+        return "redirect:/meeting/list";
     }
 
     @RequestMapping("/delete")
-    public String deleteMeeting(@RequestParam("id") Long id) {
+    public String deleteMeeting(@RequestParam("id") Long id, Authentication authentication) {
+
+        if (authentication == null) {
+            return "redirect:/member/login";
+        }
 
         Boolean deleteMeeting = meetingService.DeleteMeeting(id);
         if (deleteMeeting) {
-            return "redirect:/meeting/";
+            return "redirect:/meeting/list";
         } else {
             return "redirect:/main";
         }
